requirements.txt
Django==1.11.3
djangorestframework==3.6.3
factory-boy==2.8.1
Faker==0.7.18
psycopg2==2.7.1
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
